Rename D2Patch.h to D2BasePatch.h. - Preparation to make D2Patch the mass include file.
