+++ src/D2Patch/D2AnyPatch.cpp
@@ -29,7 +29,7 @@
 #include "../D2Offset.h"
 
 D2AnyPatch::D2AnyPatch(const D2Offset& d2Offset, const DWORD data,
-                       const bool relative, size_t patchSize) : D2Patch(d2Offset, patchSize),
+                       const bool relative, size_t patchSize) : D2BasePatch(d2Offset, patchSize),
     data(data), relative(relative) {
 }
 
@@ -41,8 +41,8 @@ D2AnyPatch::D2AnyPatch(const D2Offset& d2Offset, const OpCode opCode,
 bool D2AnyPatch::applyPatch() const {
     HANDLE gameHandle = GetCurrentProcess();
 
-    if ((getD2Offset().getCurrentOffset() & D2Patch::NO_PATCH) ==
-            D2Patch::NO_PATCH) {
+    if ((getD2Offset().getCurrentOffset() & D2BasePatch::NO_PATCH) ==
+            D2BasePatch::NO_PATCH) {
         return true;
     }
